renamed pipeline to pipelines
