--- lua/lual/pipeline/presenters/init.lua
@@ -1,7 +1,0 @@
-local presenters = {}
-
-presenters.text = require("lual.pipeline.presenters.text")
-presenters.color = require("lual.pipeline.presenters.color")
-presenters.json = require("lual.pipeline.presenters.json")
-
-return presenters
