--- lua/lual/pipeline/transformers/init.lua
@@ -1,5 +1,0 @@
-local transformers = {}
-
-transformers.noop = require("lual.pipeline.transformers.noop")
-
-return transformers
