+++ lua/lual/pipelines/outputs/syslog.lua
