+++ lua/lual/pipelines/transformers/init.lua
@@ -1,0 +1,5 @@
+local transformers = {}
+
+transformers.noop = require("lual.pipelines.transformers.noop")
+
+return transformers
