+++ spec/config/config_spec.lua
@@ -32,7 +32,7 @@ describe("lual.config() API", function()
             assert.are.equal(1, #pipelines)
             assert.is_table(pipelines[1].outputs)
             assert.are.equal(1, #pipelines[1].outputs)
-            assert_output_is_function(pipelines[1].outputs[1], lual.pipeline.outputs.console)
+            assert_output_is_function(pipelines[1].outputs[1], lual.pipelines.outputs.console)
         end)
 
         it("should preserve existing configuration for unspecified keys", function()
@@ -56,7 +56,7 @@ describe("lual.config() API", function()
             assert.are.equal(1, #pipelines)
             assert.is_table(pipelines[1].outputs)
             assert.are.equal(1, #pipelines[1].outputs)
-            assert_output_is_function(pipelines[1].outputs[1], lual.pipeline.outputs.console)
+            assert_output_is_function(pipelines[1].outputs[1], lual.pipelines.outputs.console)
         end)
 
         it("should allow updating pipelines", function()
@@ -92,7 +92,7 @@ describe("lual.config() API", function()
             assert.are.equal(1, #pipelines)
             assert.is_table(pipelines[1].outputs)
             assert.are.equal(1, #pipelines[1].outputs)
-            assert_output_is_function(pipelines[1].outputs[1], lual.pipeline.outputs.console)
+            assert_output_is_function(pipelines[1].outputs[1], lual.pipelines.outputs.console)
         end)
 
         it("should allow updating multiple keys at once", function()
@@ -322,7 +322,7 @@ describe("lual.config() API", function()
             assert.are.equal(1, #config.pipelines)
             assert.is_table(config.pipelines[1].outputs)
             assert.are.equal(1, #config.pipelines[1].outputs)
-            assert_output_is_function(config.pipelines[1].outputs[1], lual.pipeline.outputs.console)
+            assert_output_is_function(config.pipelines[1].outputs[1], lual.pipelines.outputs.console)
         end)
     end)
 
@@ -363,7 +363,7 @@ describe("lual.config() API", function()
             assert.are.equal(1, #after_reset.pipelines)
             assert.is_table(after_reset.pipelines[1].outputs)
             assert.are.equal(1, #after_reset.pipelines[1].outputs)
-            assert_output_is_function(after_reset.pipelines[1].outputs[1], lual.pipeline.outputs.console)
+            assert_output_is_function(after_reset.pipelines[1].outputs[1], lual.pipelines.outputs.console)
         end)
     end)
