+++ spec/dispatch/dispatch_spec.lua
@@ -3,10 +3,10 @@ package.path = package.path .. ";./lua/?.lua;./lua/?/init.lua;../lua/?.lua;../lu
 
 local lual = require("lual.logger")
 local core_levels = require("lua.lual.levels")
-local console = require("lual.pipeline.outputs.console")
-local file_output = require("lual.pipeline.outputs.file")
-local syslog = require("lual.pipeline.outputs.syslog")
-local all_presenters = require("lual.pipeline.presenters.init") -- For presenter tests
+local console = require("lual.pipelines.outputs.console")
+local file_output = require("lual.pipelines.outputs.file")
+local syslog = require("lual.pipelines.outputs.syslog")
+local all_presenters = require("lual.pipelines.presenters.init") -- For presenter tests
 
 -- Helper function to check if a file exists
 local function file_exists(filename)
@@ -841,7 +841,7 @@ describe("Presenter Configuration in pipelines", function()
         }
 
         -- Access the internal pipeline module
-        local pipeline_module = require("lual.pipeline")
+        local pipeline_module = require("lual.pipelines")
 
         -- Process the pipeline directly
         pipeline_module._process_pipeline(test_record, logger.pipelines[1], logger)
