+++ spec/presenters/color_presenter_spec.lua
@@ -1,8 +1,8 @@
 package.path = package.path .. ";./lua/?.lua;./lua/?/init.lua;../lua/?.lua;../lua/?/init.lua"
 local unpack = unpack or table.unpack
-local color_factory = require("lual.pipeline.presenters.color")
+local color_factory = require("lual.pipelines.presenters.color")
 
-describe("lual.pipeline.presenters.color", function()
+describe("lual.pipelines.presenters.color", function()
 	local colors = {
 		reset = "\27[0m",
 		dim = "\27[2m",
