+++ spec/presenters/presenters_spec.lua
@@ -1,9 +1,10 @@
 #!/usr/bin/env lua
 package.path = package.path .. ";./lua/?.lua;./lua/?/init.lua;../lua/?.lua;../lua/?/init.lua"
+local unpack = unpack or table.unpack
 
-local text_presenter = require("lual.pipeline.presenters.text")
-local color_presenter = require("lual.pipeline.presenters.color")
-local json_presenter = require("lual.pipeline.presenters.json")
+local text_presenter = require("lual.pipelines.presenters.text")
+local color_presenter = require("lual.pipelines.presenters.color")
+local json_presenter = require("lual.pipelines.presenters.json")
 local time_utils = require("lual.utils.time")
 
 describe("lual Presenters", function()
