+++ spec/transformers/transformers_spec.lua
@@ -1,7 +1,7 @@
 #!/usr/bin/env lua
 package.path = package.path .. ";./lua/?.lua;./lua/?/init.lua;../lua/?.lua;../lua/?/init.lua"
 
-local transformers = require("lual.pipeline.transformers")
+local transformers = require("lual.pipelines.transformers")
 
 describe("lual Transformers", function()
     -- Sample log record for testing
